features(quality) js & python for codacy
