+++ src/tipboard/app/FakeData/fakechartJS.py
@@ -1,5 +1,4 @@
-import json
-from src.tipboard.app.applicationconfig import getRedisPrefix, getIsoTime
+from src.tipboard.app.applicationconfig import getIsoTime
 
 COLOR_TAB = ['rgba(62, 149, 205, 0.8)', 'rgba(114, 191, 68, 0.8)']
 
@@ -347,167 +346,3 @@ def getFakePolarareaChart(tile_id, template_name):
         },
         'modified': getIsoTime()
     }
-
-
-def getFakeSimplePercentg(tile_id, template_name):
-    return {
-        'tile_template': template_name,
-        'meta': {
-            'big_value_color': '#4CAF50',
-            'fading_background': False
-        },
-        'data': {
-            'title': 'Title',
-            'big_value': '24',
-            'subtitle': 'Subtitle',
-            'right_value': 'right_value',
-            'right_label': 'right_label',
-            'left_label': 'left_label',
-            'left_value': 'left_value'
-        },
-        'modified': getIsoTime(),
-        'id': tile_id
-    }
-
-
-def getFakeListing(tile_id, template_name):
-    return {
-        'tile_template': template_name,
-        'meta': {
-            'fading_background': False
-        },
-        'data': {
-            'items': [
-                'Leader: 42',
-                'Product Owner: 1',
-                'Scrum Master: 1',
-                'Developer: 1',
-                'U.X: 1'
-            ]
-        },
-        'modified': getIsoTime(),
-        'id': tile_id
-    }
-
-
-def getFakeText(tile_id, template_name):
-    return {
-        'tile_template': template_name,
-        'meta': {
-            'fading_background': False
-        },
-        'data': {
-            'text': 42
-        },
-        'modified': getIsoTime(),
-        'id': tile_id
-    }
-
-
-def getFakeFancyListing(tile_id, template_name):
-    return {
-        'tile_template': template_name,
-        'meta': {
-            'big_value_color': '#4CAF50',
-            'fading_background': False
-        },
-        'data': {
-            'title': 'title',
-            'description': 'description',
-        },
-        'modified': getIsoTime(),
-        'id': tile_id
-    }
-
-
-def getFakeBigValue(tile_id, template_name):
-    return {
-        'tile_template': template_name,
-        'meta': {
-            'big_value_color': '#4CAF50',
-            'fading_background': False
-        },
-        'data': {
-            'title': 'Title',
-            'description': 'Description',
-            "big-value": "25%",
-            "lower-left-label": "lower left-label",
-            "upper-left-label": "upper left-label",
-            "upper-right-label": "upper right-label",
-            "lower-right-label": "lower right-label"
-        },
-        'modified': getIsoTime(),
-        'id': tile_id
-    }
-
-
-def getFakeJustValue(tile_id, template_name):
-    return {
-        'tile_template': template_name,
-        'meta': {
-            'big_value_color': '#4CAF50',
-            'fading_background': False
-        },
-        'data': {
-            'title': 'Title',
-            'description': 'Description',
-            'just-value': '42'
-        },
-        'modified': getIsoTime(),
-        'id': tile_id
-    }
-
-
-def buildFakeDataForChartJS(tile_id, template_name):
-    data = None
-    if template_name == 'pie_chart':
-        data = getFakePieChart(tile_id, template_name)
-    elif template_name == 'line_chart':
-        data = getFakeLineChart(tile_id, template_name)
-    elif template_name == 'cumulative_flow':
-        data = getFakeCumulFlow(tile_id, template_name)
-    elif template_name == 'bar_chart':
-        data = getFakeBarChart(tile_id, template_name)
-    elif template_name == 'norm_chart':
-        data = getFakeNormChart(tile_id, template_name)
-
-    elif template_name == 'doughnut_chart':
-        data = getFakeDoughnutChart(tile_id, template_name)
-    elif template_name == 'radar_chart':
-        data = getFakeRadarChart(tile_id, template_name)
-    elif template_name == 'polararea_chart':
-        data = getFakePolarareaChart(tile_id, template_name)
-    elif template_name == 'vbar_chart':
-        data = getFakeVbarChart(tile_id, template_name)
-    else:
-        print(f'ERROR WITH FAKE DATA ON {tile_id}', flush=True)
-    return data
-
-
-def buildDataFromValueTemplate(tile_id, template_name):
-    if template_name == 'big_value':
-        return getFakeBigValue(tile_id, template_name)
-    elif template_name == 'just_value':
-        return getFakeJustValue(tile_id, template_name)
-    elif template_name == 'simple_percentage':
-        return getFakeSimplePercentg(tile_id, template_name)
-    else:
-        return None
-
-
-def buildFakeDataFromTemplate(tile_id, template_name, cache):
-    print(f'Building fake data for {tile_id} as template: {template_name}')
-    data = dict()
-    if template_name == 'text':
-        data = getFakeText(tile_id, template_name)
-    elif template_name == 'listing':
-        data = getFakeListing(tile_id, template_name)
-    elif template_name == 'fancy_listing':
-        data = getFakeFancyListing(tile_id, template_name)
-    elif 'value' in template_name or 'percentage' in template_name:
-        data = buildDataFromValueTemplate(tile_id, template_name)
-    elif template_name != 'empty':
-        data = buildFakeDataForChartJS(tile_id, template_name)
-    if cache is not None:
-        cache.redis.set(name=getRedisPrefix(tile_id), value=json.dumps(data))
-    return data
